Fix load after click logic so that link buttons work again
When I changed it so that it only happened when forms were valid, it
broke all the buttons that weren't forms. Added logic that makes all
links (non-form) with .load-after-click disable and show loading icon
when clicked. This doesn't affect the form buttons because form buttons
are <button> and this only affects <a>.